+++ Assets/LKH/Materials/PlayerBullet.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 96002dab60c8f17498ba354fcbff6f28
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
