+++ Assets/LKH/Prefabs/Player.prefab
@@ -13,15 +13,10 @@ GameObject:
   - component: {fileID: 2164884307889646682}
   - component: {fileID: 6845309445483451645}
   - component: {fileID: 8381243986527022444}
-  - component: {fileID: 2697088649480490682}
-  - component: {fileID: 7799013040172943913}
-  - component: {fileID: 8107589362572569315}
-  - component: {fileID: 8438963447911530627}
-  - component: {fileID: 3937761547066315206}
   - component: {fileID: 6956037539318070097}
   m_Layer: 0
   m_Name: Player
-  m_TagString: Untagged
+  m_TagString: Player
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
@@ -35,7 +30,7 @@ Transform:
   m_GameObject: {fileID: 6669484866285966142}
   serializedVersion: 2
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalPosition: {x: 0, y: 1, z: 0}
   m_LocalScale: {x: 2, y: 2, z: 2}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -99,9 +94,9 @@ Rigidbody:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 6669484866285966142}
   serializedVersion: 4
-  m_Mass: 1
-  m_Drag: 5
-  m_AngularDrag: 4
+  m_Mass: 3
+  m_Drag: 0
+  m_AngularDrag: 0
   m_CenterOfMass: {x: 0, y: 0, z: 0}
   m_InertiaTensor: {x: 1, y: 1, z: 1}
   m_InertiaRotation: {x: 0, y: 0, z: 0, w: 1}
@@ -139,73 +134,6 @@ SphereCollider:
   serializedVersion: 3
   m_Radius: 0.5
   m_Center: {x: 0, y: 0, z: 0}
---- !u!114 &2697088649480490682
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6669484866285966142}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 756785dec4f1645f687e1162f4b2db95, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
---- !u!114 &7799013040172943913
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6669484866285966142}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 66128c0a304da8e4c900eef6875fe504, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  speed: 10
---- !u!114 &8107589362572569315
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6669484866285966142}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 4c01d3e102714c249a713eb98544aeb8, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  cube: {fileID: 0}
-  initBallSpeed: 0
-  m_initBallVector: {x: 0, y: 0, z: 0}
---- !u!114 &8438963447911530627
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6669484866285966142}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f66d82faffbb33a4d99e926c5c42328d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  ballSpeed: 0
-  ballMovementVector: {x: 0, y: 0, z: 0}
---- !u!114 &3937761547066315206
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6669484866285966142}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b017063ee8eefa0468751ecbd2b83eb2, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  gravityScale: 5
 --- !u!114 &6956037539318070097
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -218,6 +146,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: eb28778401aa04afcbb181c09a0cf996, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  jumpPower: 0
+  jumpPower: 10
   ItemCount: 0
   manager: {fileID: 0}
