+++ Assets/LKH/Prefabs/PlayerBullet.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d1ad14509d0fb344ba94912c8c09190f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
