+++ Assets/LKH/Scripts/PlayerSpawner.cs
@@ -1,0 +1,13 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PlayerSpawner : MonoBehaviour
+{
+    [SerializeField] GameObject playerPrefab;
+    private void Start()
+    {
+        GameObject player = Instantiate(playerPrefab);
+        player.transform.position = transform.position;
+    }
+}
