+++ ProjectSettings/TagManager.asset
@@ -7,6 +7,7 @@ TagManager:
   - Item
   - Monster
   - Floor
+  - PlayerBullet
   layers:
   - Default
   - TransparentFX
@@ -16,7 +17,7 @@ TagManager:
   - UI
   - Wall
   - Obstacle
-  - 
+  - PlayerBullet
   - 
   - 
   -
